fixed card alignmet issue
