+++ src/components/Button/Style.js
@@ -2,7 +2,7 @@ import { Button } from '@material-ui/core';
 import styled from 'styled-components';
 
 const StyledButton = styled(Button)`
-  font-size: 20px;
+  font-size: 10px;
   letter-spacing: 2px;
   margin-bottom: 10px;
   max-width: ${({ maxwidth }) => maxwidth};
