+++ src/components/CardMenus/Tabs.js
@@ -1,32 +1,23 @@
-import React from 'react';
+import React, { useState } from 'react';
 
 import { Paper, Tab, Tabs } from '@material-ui/core';
 
 import CardMenus from './CardMenus';
 
 const MainTab = () => {
-  const [value, setValue] = React.useState(2);
-
+  const [value, setValue] = useState(2);
   const handleChange = (event, newValue) => {
     setValue(newValue);
   };
   return (
     <>
       <Paper square>
-        <Tabs
-          aria-label="disabled tabs example"
-          indicatorColor="primary"
-          onChange={handleChange}
-          textColor="primary"
-          value={value}
-        >
+        <Tabs indicatorColor="secondary" onChange={handleChange} textColor="secondary" value={value}>
           <Tab label="Gravy" />
           <Tab label="Bread" />
           <Tab label="Rice" />
           <Tab label="Burgers" />
           <Tab label="Rolls" />
-          <Tab label="Pizza" />
-          <Tab label="Chinese" />
         </Tabs>
       </Paper>
       <CardMenus />
